phase2/pyproject.toml
[project]
name = "labs-rqaoa"
version = "0.1.0"
description = "Recursive QAOA for LABS Problem - iQuHack 2026 NVIDIA Challenge"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x>=12.0.0",
]
dev = [
    "pytest>=7.0.0",
    "hypothesis>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/utils", "src/solvers", "src/algorithms", "benchmarks"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
